ProperNounFinder/requirements.txt
openai==1.35.15
pandas==2.1.4
python-dotenv==1.0.0 